requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
dj-database-url==0.5.0
Django==3.1.7
django-3-jet==1.0.8
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-extensions==3.1.1
django-heroku==0.3.1
django-js-asset==1.2.2
graphviz==0.16
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
pydotplus==2.0.2
PyJWT==2.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
